UI tests open IntelliJ settings on mac
The SettingsComponentTest UI test suite is now able to open IntelliJ settings on macOS. Several if statements were added to enable this functionality.
One of the assertions had to be changed for the test to pass.

It was necessary to comment out @Video because it caused UI tests to fail. This could be a problem of my computer, but this remains to be tested.