General Refactor (#440)
* Move class-level fields in TestSparkAction to actionPerformed

Because this may lead to resource leaks. Closes #242

* Remove redundant qualifier names in TestSparkAction.kt

* Extract the TestSparkActionWindow class into a separate file

Extract the TestSparkActionWindow class into a separate file for better modularity and maintainability. This separation improves code readability and simplifies the structure of TestSparkAction by isolating UI logic. It will also reduce potential git conflicts.

* Add intelliJPlatform folder to gitignore

* Remove the irrelevant TODO in TestSparkActionWindow.kt

If the psiHelper is not available, we'll get a NullPointerException automatically, so we don't need to throw anything manually. It's the responsibility of the caller class to make sure that the window is shown only when the dataContext returns some data and a psiHelper instance can be retrieved.

* Refactor TestCompilerFactory to resolve issue #324

Use the provided javaSDKHomePath or derive it from the project's instance. If can't derive throw an exception.

* Resolve ktlint issues